backend/requirements.txt
# Core dependencies for Agents Marketplace Backend
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0

# Development dependencies (optional)
pytest==7.4.3
black==23.11.0
flake8==6.1.0
